requirements.txt
certifi==2024.8.30
charset-normalizer==3.4.0
contourpy==1.3.0
cycler==0.12.1
faiss-cpu==1.9.0
filelock==3.16.1
fonttools==4.54.1
fsspec==2024.9.0
huggingface-hub==0.25.2
idna==3.10
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.7
MarkupSafe==3.0.1
matplotlib==3.9.2
mpmath==1.3.0
networkx==3.4.1
numpy==2.1.2
packaging==24.1
pillow==10.4.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
PyYAML==6.0.2
regex==2024.9.11
requests==2.32.3
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
sentence-transformers==3.2.0
six==1.16.0
sympy==1.13.3
threadpoolctl==3.5.0
tokenizers==0.20.1
torch==2.4.1
tqdm==4.66.5
transformers==4.45.2
typing_extensions==4.12.2
urllib3==2.2.3
